// types.ts
export interface Trait {
  label: string;
  description: string;
  icon?: string;
}

export interface ProfileData {
  name: string;
  age: number;
  job: string;
  height: string;
  weight_lift: string;
  mbti: string;
  visual: string;
  location: string;
}

export interface PrologueData {
  date: string;
  situation: string;
  summary: string;
}